Module arrow_array::array

The concrete array definitions

Structs

BooleanArray

Array of bools

DictionaryArray

A dictionary array where each element is a single value indexed by an integer key. This is mostly used to represent strings or a limited set of primitive types as integers, for example when doing NLP analysis or representing chromosomes by name.

FixedSizeBinaryArray

An array where each element is a fixed-size sequence of bytes.

FixedSizeListArray

A list array where each element is a fixed-size sequence of values with the same type whose maximum length is represented by a i32.

GenericByteArray

Generic struct for variable-size byte arrays

GenericListArray

Generic struct for a variable-size list array.

MapArray

A nested array type where each record is a key-value map. Keys should always be non-null, but values can be null.

NativeAdapter

This struct is used as an adapter when creating PrimitiveArray from an iterator. FromIterator for PrimitiveArray takes an iterator where the elements can be into this struct. So once implementing From or Into trait for a type, an iterator of the type can be collected to PrimitiveArray.

NullArray

An Array where all elements are nulls

PrimitiveArray

Array whose elements are of primitive types.

RunArray

A run-end encoding (REE) is a variation of run-length encoding (RLE).

StructArray

A nested array type where each child (called field) is represented by a separate array.

TypedDictionaryArray

A strongly-typed wrapper around a DictionaryArray that implements ArrayAccessor allowing fast access to its elements

UnionArray

An Array that can represent slots of varying types.

Traits

Array

Trait for dealing with different types of array at runtime when the type of the array is not known in advance.

ArrayAccessor

A generic trait for accessing the values of an Array

ArrowPrimitiveType

Trait bridging the dynamic-typed nature of Arrow (via DataType) with the static-typed nature of rust types (ArrowNativeType) for all types that implement ArrowNativeType.

OffsetSizeTrait

trait declaring an offset size, relevant for i32 vs i64 array types.

Functions

make_array

Constructs an array using the input data. Returns a reference-counted Array instance.

new_empty_array

Creates a new empty array

new_null_array

Creates a new array of data_type of length length filled entirely of NULL values

Type Definitions

ArrayRef

A reference-counted reference to a generic Array.

BinaryArray

An array where each element contains 0 or more bytes. The byte length of each element is represented by an i32.

Date32Array

A primitive array where each element is of 32-bit date type.

Date64Array

A primitive array where each element is of 64-bit date type.

Decimal128Array

An array where each element is a 128-bits decimal with precision in [1, 38] and scale less or equal to 38.

Decimal256Array

An array where each element is a 256-bits decimal with precision in [1, 76] and scale less or equal to 76.

DurationMicrosecondArray

An array where each element is an elapsed time type in microseconds.

DurationMillisecondArray

An array where each element is an elapsed time type in milliseconds.

DurationNanosecondArray

An array where each element is an elapsed time type in nanoseconds.

DurationSecondArray

An array where each element is an elapsed time type in seconds.

Float16Array

Example: Using collect

Float32Array

Example: Using collect

Float64Array

Example: Using collect

GenericBinaryArray

See BinaryArray and LargeBinaryArray for storing binary data.

GenericStringArray

Generic struct for [Large]StringArray

Int8Array

Example: Using collect

Int8DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

Int16Array

Example: Using collect

Int16DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

Int16RunArray

A RunArray array where run ends are stored using i16 data type.

Int32Array

Example: Using collect

Int32DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

Int32RunArray

A RunArray array where run ends are stored using i32 data type.

Int64Array

Example: Using collect

Int64DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

Int64RunArray

A RunArray array where run ends are stored using i64 data type.

IntervalDayTimeArray

An array where each element is a “calendar” interval days and milliseconds.

IntervalMonthDayNanoArray

An array where each element is a “calendar” interval in months, days, and nanoseconds.

IntervalYearMonthArray

An array where each element is a “calendar” interval in months.

LargeBinaryArray

An array where each element contains 0 or more bytes. The byte length of each element is represented by an i64.

LargeListArray

A list array where each element is a variable-sized sequence of values with the same type whose memory offsets between elements are represented by a i64.

LargeStringArray

An array where each element is a variable-sized sequence of bytes representing a string whose maximum length (in bytes) is represented by a i64.

ListArray

A list array where each element is a variable-sized sequence of values with the same type whose memory offsets between elements are represented by a i32.

StringArray

An array where each element is a variable-sized sequence of bytes representing a string whose maximum length (in bytes) is represented by a i32.

Time32MillisecondArray

An array where each element is of 32-bit type representing time elapsed in milliseconds since midnight.

Time32SecondArray

An array where each element is of 32-bit type representing time elapsed in seconds since midnight.

Time64MicrosecondArray

An array where each element is of 64-bit type representing time elapsed in microseconds since midnight.

Time64NanosecondArray

An array where each element is of 64-bit type representing time elapsed in nanoseconds since midnight.

TimestampMicrosecondArray

A primitive array where each element is of type TimestampMicrosecondType. See examples for TimestampSecondArray.

TimestampMillisecondArray

A primitive array where each element is of type TimestampMillisecondType. See examples for TimestampSecondArray.

TimestampNanosecondArray

A primitive array where each element is of type TimestampNanosecondType. See examples for TimestampSecondArray.

TimestampSecondArray

A primitive array where each element is of type TimestampSecondType. See also Timestamp.

UInt8Array

Example: Using collect

UInt8DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

UInt16Array

Example: Using collect

UInt16DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

UInt32Array

Example: Using collect

UInt32DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.

UInt64Array

Example: Using collect

UInt64DictionaryArray

A dictionary array where each element is a single value indexed by an integer key.