Type Alias arrow_array::array::list_array::LargeListArray

pub type LargeListArray = GenericListArray<i64>;

A GenericListArray of variable size lists, storing offsets as i64.

Aliased Type

struct LargeListArray {
    data_type: DataType,
    nulls: Option<NullBuffer>,
    values: Arc<dyn Array, Global>,
    value_offsets: OffsetBuffer<i64>,
}

Fields
data_type: DataTypenulls: Option<NullBuffer>values: Arc<dyn Array, Global>value_offsets: OffsetBuffer<i64>
Implementations

impl<OffsetSize: OffsetSizeTrait> GenericListArray<OffsetSize>

pub const DATA_TYPE_CONSTRUCTOR: fn(_: FieldRef) -> DataType = _

The data type constructor of list array. The input is the schema of the child array and the output is the [DataType], List or LargeList.

pub fn try_new( field: FieldRef, offsets: OffsetBuffer<OffsetSize>, values: ArrayRef, nulls: Option<NullBuffer> ) -> Result<Self, ArrowError>

Create a new GenericListArray from the provided parts

Errors

Errors if

offsets.len() - 1 != nulls.len()

offsets.last() > values.len()

!field.is_nullable() && values.is_nullable()

field.data_type() != values.data_type()

pub fn new( field: FieldRef, offsets: OffsetBuffer<OffsetSize>, values: ArrayRef, nulls: Option<NullBuffer> ) -> Self

Create a new GenericListArray from the provided parts

Panics

Panics if Self::try_new returns an error

pub fn new_null(field: FieldRef, len: usize) -> Self

Create a new GenericListArray of length len where all values are null

pub fn into_parts( self ) -> (FieldRef, OffsetBuffer<OffsetSize>, ArrayRef, Option<NullBuffer>)

Deconstruct this array into its constituent parts

pub fn offsets(&self) -> &OffsetBuffer<OffsetSize>

Returns a reference to the offsets of this list

Unlike Self::value_offsets this returns the [OffsetBuffer] allowing for zero-copy cloning

pub fn values(&self) -> &ArrayRef

Returns a reference to the values of this list

pub fn value_type(&self) -> DataType

Returns a clone of the value type of this list.

pub unsafe fn value_unchecked(&self, i: usize) -> ArrayRef

Returns ith value of this list array.

Safety

Caller must ensure that the index is within the array bounds

pub fn value(&self, i: usize) -> ArrayRef

Returns ith value of this list array.

pub fn value_offsets(&self) -> &[OffsetSize]

Returns the offset values in the offsets buffer

pub fn value_length(&self, i: usize) -> OffsetSize

Returns the length for value at index i.

pub fn iter<'a>(&'a self) -> GenericListArrayIter<'a, OffsetSize>

constructs a new iterator

fn get_type(data_type: &DataType) -> Option<&DataType>

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a zero-copy slice of this array with the indicated offset and length.

pub fn from_iter_primitive<T, P, I>(iter: I) -> Selfwhere T: ArrowPrimitiveType, P: IntoIterator<Item = Option<<T as ArrowPrimitiveType>::Native>>, I: IntoIterator<Item = Option<P>>,

Creates a GenericListArray from an iterator of primitive values

Example

let data = vec![
   Some(vec![Some(0), Some(1), Some(2)]),
   None,
   Some(vec![Some(3), None, Some(5)]),
   Some(vec![Some(6), Some(7)]),
];
let list_array = ListArray::from_iter_primitive::<Int32Type, _, _>(data);
println!("{:?}", list_array);

impl<OffsetSize: OffsetSizeTrait> GenericListArray<OffsetSize>

fn try_new_from_array_data(data: ArrayData) -> Result<Self, ArrowError>