arrow_array::array::run_array

Trait AnyRunEndArray

pub trait AnyRunEndArray: Array {
    // Required methods
    fn values(&self) -> &Arc<dyn Array> ;
    fn with_values(&self, values: ArrayRef) -> ArrayRef;
}

An array that can be downcast to a RunArray of any run end type and any value type.

This can be used to efficiently implement kernels for all possible run end types without needing to create specialized implementations for each key type.

Required Methods

fn values(&self) -> &Arc<dyn Array>

Returns the values of this array.

fn with_values(&self, values: ArrayRef) -> ArrayRef

Returns a new run-end encoded array with the given values, preserving the existing run ends.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".