Struct arrow_array::array::FixedSizeListArray

pub struct FixedSizeListArray { /* private fields */ }

A list array where each element is a fixed-size sequence of values with the same type whose maximum length is represented by a i32.

Example

// Construct a value array
let value_data = ArrayData::builder(DataType::Int32)
    .len(9)
    .add_buffer(Buffer::from_slice_ref(&[0, 1, 2, 3, 4, 5, 6, 7, 8]))
    .build()
    .unwrap();
let list_data_type = DataType::FixedSizeList(
    Box::new(Field::new("item", DataType::Int32, false)),
    3,
);
let list_data = ArrayData::builder(list_data_type.clone())
    .len(3)
    .add_child_data(value_data.clone())
    .build()
    .unwrap();
let list_array = FixedSizeListArray::from(list_data);
let list0 = list_array.value(0);
let list1 = list_array.value(1);
let list2 = list_array.value(2);

assert_eq!( &[0, 1, 2], list0.as_any().downcast_ref::<Int32Array>().unwrap().values());
assert_eq!( &[3, 4, 5], list1.as_any().downcast_ref::<Int32Array>().unwrap().values());
assert_eq!( &[6, 7, 8], list2.as_any().downcast_ref::<Int32Array>().unwrap().values());

For non generic lists, you may wish to consider using crate::array::FixedSizeBinaryArray

Implementations

impl FixedSizeListArray

pub fn values(&self) -> ArrayRef

Returns a reference to the values of this list.

pub fn value_type(&self) -> DataType

Returns a clone of the value type of this list.

pub fn value(&self, i: usize) -> ArrayRef

Returns ith value of this list array.

pub fn value_offset(&self, i: usize) -> i32

Returns the offset for value at index i.

Note this doesn’t do any bound checking, for performance reason.

pub const fn value_length(&self) -> i32

Returns the length for an element.

All elements have the same length as the array is a fixed size.

pub fn from_iter_primitive<T, P, I>(iter: I, length: i32) -> Selfwhere T: ArrowPrimitiveType, P: IntoIterator<Item = Option<<T as ArrowPrimitiveType>::Native>>, I: IntoIterator<Item = Option<P>>,

Creates a FixedSizeListArray from an iterator of primitive values

Example

let data = vec![
   Some(vec![Some(0), Some(1), Some(2)]),
   None,
   Some(vec![Some(3), None, Some(5)]),
   Some(vec![Some(6), Some(7), Some(45)]),
];
let list_array = FixedSizeListArray::from_iter_primitive::<Int32Type, _, _>(data, 3);
println!("{:?}", list_array);