Struct arrow_array::array::GenericByteViewArray

pub struct GenericByteViewArray<T: ByteViewType + ?Sized> {
    data_type: DataType,
    views: ScalarBuffer<u128>,
    buffers: Vec<Buffer>,
    phantom: PhantomData<T>,
    nulls: Option<NullBuffer>,
}

Variable-size Binary View Layout: An array of variable length bytes view arrays.

Different than crate::GenericByteArray as it stores both an offset and length meaning that take / filter operations can be implemented without copying the underlying data.

See StringViewArray for storing utf8 encoded string data and BinaryViewArray for storing bytes.

A GenericByteViewArray stores variable length byte strings. An array of N elements is stored as N fixed length “views” and a variable number of variable length “buffers”.

Each view is a u128 value layout is different depending on the length of the string stored at that location:

                        ┌──────┬────────────────────────┐
                        │length│      string value      │
   Strings (len <= 12)  │      │    (padded with 0)     │
                        └──────┴────────────────────────┘
                         0    31                      127

                        ┌───────┬───────┬───────┬───────┐
                        │length │prefix │  buf  │offset │
   Strings (len > 12)   │       │       │ index │       │
                        └───────┴───────┴───────┴───────┘
                         0    31       63      95    127

Strings with length <= 12 are stored directly in the view.

Strings with length > 12: The first four bytes are stored inline in the view and the entire string is stored in one of the buffers.

Unlike GenericByteArray, there are no constraints on the offsets other than they must point into a valid buffer. However, they can be out of order, non continuous and overlapping.

For example, in the following diagram, the strings “FishWasInTownToday” and “CrumpleFacedFish” are both longer than 12 bytes and thus are stored in a separate buffer while the string “LavaMonster” is stored inlined in the view. In this case, the same bytes for “Fish” are used to store both strings.

                                                                           ┌───┐
                        ┌──────┬──────┬──────┬──────┐               offset │...│
"FishWasInTownTodayYay" │  21  │ Fish │  0   │ 115  │─ ─              103  │Mr.│
                        └──────┴──────┴──────┴──────┘   │      ┌ ─ ─ ─ ─ ▶ │Cru│
                        ┌──────┬──────┬──────┬──────┐                      │mpl│
"CrumpleFacedFish"      │  16  │ Crum │  0   │ 103  │─ ─│─ ─ ─ ┘           │eFa│
                        └──────┴──────┴──────┴──────┘                      │ced│
                        ┌──────┬────────────────────┐   └ ─ ─ ─ ─ ─ ─ ─ ─ ▶│Fis│
"LavaMonster"           │  11  │   LavaMonster\0    │                      │hWa│
                        └──────┴────────────────────┘               offset │sIn│
                                                                      115  │Tow│
                                                                           │nTo│
                                                                           │day│
                                 u128 "views"                              │Yay│
                                                                  buffer 0 │...│
                                                                           └───┘

Fields
data_type: DataTypeviews: ScalarBuffer<u128>buffers: Vec<Buffer>phantom: PhantomData<T>nulls: Option<NullBuffer>
Implementations

impl<T: ByteViewType + ?Sized> GenericByteViewArray<T>

pub fn new( views: ScalarBuffer<u128>, buffers: Vec<Buffer>, nulls: Option<NullBuffer>, ) -> Self

Create a new GenericByteViewArray from the provided parts, panicking on failure

Panics

Panics if GenericByteViewArray::try_new returns an error

pub fn try_new( views: ScalarBuffer<u128>, buffers: Vec<Buffer>, nulls: Option<NullBuffer>, ) -> Result<Self, ArrowError>

Create a new GenericByteViewArray from the provided parts, returning an error on failure

Errors

views.len() != nulls.len()

ByteViewType::validate fails

pub unsafe fn new_unchecked( views: ScalarBuffer<u128>, buffers: Vec<Buffer>, nulls: Option<NullBuffer>, ) -> Self

Create a new GenericByteViewArray from the provided parts, without validation

Safety

Safe if Self::try_new would not error

pub fn new_null(len: usize) -> Self

Create a new GenericByteViewArray of length len where all values are null

pub fn new_scalar(value: impl AsRef<T::Native>) -> Scalar<Self>

Create a new Scalar from value

pub fn from_iter_values<Ptr, I>(iter: I) -> Self
where Ptr: AsRef<T::Native>, I: IntoIterator<Item = Ptr>,

Creates a GenericByteViewArray based on an iterator of values without nulls

pub fn into_parts(self) -> (ScalarBuffer<u128>, Vec<Buffer>, Option<NullBuffer>)

Deconstruct this array into its constituent parts

pub fn views(&self) -> &ScalarBuffer<u128>

Returns the views buffer

pub fn data_buffers(&self) -> &[Buffer]

Returns the buffers storing string data

pub fn value(&self, i: usize) -> &T::Native

Returns the element at index i

Panics

Panics if index i is out of bounds.

pub unsafe fn value_unchecked(&self, idx: usize) -> &T::Native

Returns the element at index i

Safety

Caller is responsible for ensuring that the index is within the bounds of the array

pub unsafe fn inline_value(view: &u128, len: usize) -> &[u8]

Returns the inline value of the view.

Safety

The view must be a valid element from Self::views() that adheres to the view layout.

The len must be the length of the inlined value. It should never be larger than 12.

pub fn iter(&self) -> ArrayIter<&Self>

constructs a new iterator

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a zero-copy slice of this array with the indicated offset and length.

pub fn gc(&self) -> Self

Returns a “compacted” version of this array

The original array will not be modified

Garbage Collection

Before GC:

                                       ┌──────┐                 
                                       │......│                 
                                       │......│                 
┌────────────────────┐       ┌ ─ ─ ─ ▶ │Data1 │   Large buffer  
│       View 1       │─ ─ ─ ─          │......│  with data that
├────────────────────┤                 │......│ is not referred
│       View 2       │─ ─ ─ ─ ─ ─ ─ ─▶ │Data2 │ to by View 1 or
└────────────────────┘                 │......│      View 2     
                                       │......│                 
   2 views, refer to                   │......│                 
  small portions of a                  └──────┘                 
     large buffer                                               

After GC:

┌────────────────────┐                 ┌─────┐    After gc, only
│       View 1       │─ ─ ─ ─ ─ ─ ─ ─▶ │Data1│     data that is  
├────────────────────┤       ┌ ─ ─ ─ ▶ │Data2│    pointed to by  
│       View 2       │─ ─ ─ ─          └─────┘     the views is  
└────────────────────┘                                 left      
                                                                  
                                                                  
        2 views                                                  

This method will compact the data buffers by recreating the view array and only include the data that is pointed to by the views.

Note that it will copy the array regardless of whether the original array is compact. Use with caution as this can be an expensive operation, only use it when you are sure that the view array is significantly smaller than when it is originally created, e.g., after filtering or slicing.

impl GenericByteViewArray<BinaryViewType>

pub fn to_string_view(self) -> Result<StringViewArray, ArrowError>

Convert the BinaryViewArray to StringViewArray If items not utf8 data, validate will fail and error returned.

pub unsafe fn to_string_view_unchecked(self) -> StringViewArray

Convert the BinaryViewArray to StringViewArray

Safety

Caller is responsible for ensuring that items in array are utf8 data.

impl GenericByteViewArray<StringViewType>

pub fn to_binary_view(self) -> BinaryViewArray

Convert the StringViewArray to BinaryViewArray

pub fn is_ascii(&self) -> bool

Returns true if all data within this array is ASCII