Struct arrow_array::array::RunArray

pub struct RunArray<R: RunEndIndexType> { /* private fields */ }

A run-end encoding (REE) is a variation of run-length encoding (RLE).

This encoding is good for representing data containing same values repeated consecutively.

RunArray contains run_ends array and values array of same length. The run_ends array stores the indexes at which the run ends. The values array stores the value of each run. Below example illustrates how a logical array is represented in RunArray

┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─┐
  ┌─────────────────┐  ┌─────────┐       ┌─────────────────┐
│ │        A        │  │    2    │ │     │        A        │     
  ├─────────────────┤  ├─────────┤       ├─────────────────┤
│ │        D        │  │    3    │ │     │        A        │    run length of 'A' = runs_ends[0] - 0 = 2
  ├─────────────────┤  ├─────────┤       ├─────────────────┤
│ │        B        │  │    6    │ │     │        D        │    run length of 'D' = run_ends[1] - run_ends[0] = 1
  └─────────────────┘  └─────────┘       ├─────────────────┤
│        values          run_ends  │     │        B        │     
                                         ├─────────────────┤
└ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─┘     │        B        │     
                                         ├─────────────────┤
               RunArray                  │        B        │    run length of 'B' = run_ends[2] - run_ends[1] = 3
              length = 3                 └─────────────────┘
  
                                            Logical array
                                               Contents

Implementations

impl<R: RunEndIndexType> RunArray<R>

pub fn try_new( run_ends: &PrimitiveArray<R>, values: &dyn Array ) -> Result<Self, ArrowError>

Attempts to create RunArray using given run_ends (index where a run ends) and the values (value of the run). Returns an error if the given data is not compatible with RunEndEncoded specification.

pub fn run_ends(&self) -> &PrimitiveArray<R>

Returns a reference to run_ends array

Note: any slicing of this array is not applied to the returned array and must be handled separately

pub fn values(&self) -> &ArrayRef

Returns a reference to values array