arrow_array::array

Trait StringArrayType

pub trait StringArrayType<'a>: ArrayAccessor<Item = &'a str> + Sized {
    // Required methods
    fn is_ascii(&self) -> bool;
    fn iter(&self) -> ArrayIter<Self> ;
}

A trait for Arrow String Arrays, currently three types are supported:

StringArray

LargeStringArray

StringViewArray

This trait helps to abstract over the different types of string arrays so that we don’t need to duplicate the implementation for each type.

Required Methods

fn is_ascii(&self) -> bool

Returns true if all data within this string array is ASCII

fn iter(&self) -> ArrayIter<Self>

Constructs a new iterator

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.