Type Alias arrow_array::array::Int64RunArray

pub type Int64RunArray = RunArray<Int64Type>;

A RunArray with i64 run ends

Example: Using collect

let array: Int64RunArray = vec!["a", "a", "b", "c", "c"].into_iter().collect();
let values: Arc<dyn Array> = Arc::new(StringArray::from(vec!["a", "b", "c"]));
assert_eq!(array.run_ends().values(), &[2, 3, 5]);
assert_eq!(array.values(), &values);

Aliased Type

struct Int64RunArray {
    data_type: DataType,
    run_ends: RunEndBuffer<i64>,
    values: Arc<dyn Array, Global>,
}

Fields
data_type: DataTyperun_ends: RunEndBuffer<i64>values: Arc<dyn Array, Global>
Implementations

impl<R: RunEndIndexType> RunArray<R>

pub fn logical_len(run_ends: &PrimitiveArray<R>) -> usize

Calculates the logical length of the array encoded by the given run_ends array.

pub fn try_new( run_ends: &PrimitiveArray<R>, values: &dyn Array ) -> Result<Self, ArrowError>

Attempts to create RunArray using given run_ends (index where a run ends) and the values (value of the run). Returns an error if the given data is not compatible with RunEndEncoded specification.

pub fn run_ends(&self) -> &RunEndBuffer<R::Native>

Returns a reference to [RunEndBuffer]

pub fn values(&self) -> &ArrayRef

Returns a reference to values array

Note: any slicing of this RunArray array is not applied to the returned array and must be handled separately

pub fn get_start_physical_index(&self) -> usize

Returns the physical index at which the array slice starts.

pub fn get_end_physical_index(&self) -> usize

Returns the physical index at which the array slice ends.

pub fn downcast<V: 'static>(&self) -> Option<TypedRunArray<'_, R, V>>

Downcast this RunArray to a TypedRunArray

use arrow_array::{Array, ArrayAccessor, RunArray, StringArray, types::Int32Type};

let orig = [Some("a"), Some("b"), None];
let run_array = RunArray::<Int32Type>::from_iter(orig);
let typed = run_array.downcast::<StringArray>().unwrap();
assert_eq!(typed.value(0), "a");
assert_eq!(typed.value(1), "b");
assert!(typed.values().is_null(2));

pub fn get_physical_index(&self, logical_index: usize) -> usize

Returns index to the physical array for the given index to the logical array. This function adjusts the input logical index based on ArrayData::offset Performs a binary search on the run_ends array for the input index.

The result is arbitrary if logical_index >= self.len()

pub fn get_physical_indices<I>( &self, logical_indices: &[I] ) -> Result<Vec<usize>, ArrowError>where I: ArrowNativeType,

Returns the physical indices of the input logical indices. Returns error if any of the logical index cannot be converted to physical index. The logical indices are sorted and iterated along with run_ends array to find matching physical index. The approach used here was chosen over finding physical index for each logical index using binary search using the function get_physical_index. Running benchmarks on both approaches showed that the approach used here scaled well for larger inputs. See https://github.com/apache/arrow-rs/pull/3622#issuecomment-1407753727 for more details.

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a zero-copy slice of this array with the indicated offset and length.