Type Alias arrow_array::array::Int8Array

pub type Int8Array = PrimitiveArray<Int8Type>;

A PrimitiveArray of i8

Examples

Construction

// Create from Vec<Option<i8>>
let arr = Int8Array::from(vec![Some(1), None, Some(2)]);
// Create from Vec<i8>
let arr = Int8Array::from(vec![1, 2, 3]);
// Create iter/collect
let arr: Int8Array = std::iter::repeat(42).take(10).collect();

See PrimitiveArray for more information and examples

Aliased Type

struct Int8Array {
    data_type: DataType,
    values: ScalarBuffer<i8>,
    nulls: Option<NullBuffer>,
}

Fields
data_type: DataTypevalues: ScalarBuffer<i8>

Values data

nulls: Option<NullBuffer>
Implementations

impl<T: ArrowPrimitiveType> PrimitiveArray<T>

pub fn new(values: ScalarBuffer<T::Native>, nulls: Option<NullBuffer>) -> Self

Create a new PrimitiveArray from the provided values and nulls

Panics

Panics if Self::try_new returns an error

Example

Creating a PrimitiveArray directly from a [ScalarBuffer] and [NullBuffer] using this constructor is the most performant approach, avoiding any additional allocations

// [1, 2, 3, 4]
let array = Int32Array::new(vec![1, 2, 3, 4].into(), None);
// [1, null, 3, 4]
let nulls = NullBuffer::from(vec![true, false, true, true]);
let array = Int32Array::new(vec![1, 2, 3, 4].into(), Some(nulls));

pub fn new_null(length: usize) -> Self

Create a new PrimitiveArray of the given length where all values are null

pub fn try_new( values: ScalarBuffer<T::Native>, nulls: Option<NullBuffer> ) -> Result<Self, ArrowError>

Create a new PrimitiveArray from the provided values and nulls

Errors

Errors if:

values.len() != nulls.len()

pub fn new_scalar(value: T::Native) -> Scalar<Self>

Create a new Scalar from value

pub fn into_parts( self ) -> (DataType, ScalarBuffer<T::Native>, Option<NullBuffer>)

Deconstruct this array into its constituent parts

pub fn with_data_type(self, data_type: DataType) -> Self

Overrides the [DataType] of this PrimitiveArray

Prefer using Self::with_timezone or Self::with_precision_and_scale where the primitive type is suitably constrained, as these cannot panic

Panics

Panics if ![Self::is_compatible]

fn assert_compatible(data_type: &DataType)

Asserts that data_type is compatible with Self

pub fn len(&self) -> usize

Returns the length of this array.

pub fn is_empty(&self) -> bool

Returns whether this array is empty.

pub fn values(&self) -> &ScalarBuffer<T::Native>

Returns the values of this array

pub fn builder(capacity: usize) -> PrimitiveBuilder<T>

Returns a new primitive array builder

pub fn is_compatible(data_type: &DataType) -> bool

Returns if this PrimitiveArray is compatible with the provided [DataType]

This is equivalent to data_type == T::DATA_TYPE, however ignores timestamp timezones and decimal precision and scale

pub unsafe fn value_unchecked(&self, i: usize) -> T::Native

Returns the primitive value at index i.

Safety

caller must ensure that the passed in offset is less than the array len()

pub fn value(&self, i: usize) -> T::Native

Returns the primitive value at index i.

Panics

Panics if index i is out of bounds

pub fn from_iter_values<I: IntoIterator<Item = T::Native>>(iter: I) -> Self

Creates a PrimitiveArray based on an iterator of values without nulls

pub fn from_value(value: T::Native, count: usize) -> Self

Creates a PrimitiveArray based on a constant value with count elements

pub fn take_iter<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<T::Native>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<T::Native>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the offsets in the iterator are less than the array len()

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a zero-copy slice of this array with the indicated offset and length.

pub fn reinterpret_cast<K>(&self) -> PrimitiveArray<K>where K: ArrowPrimitiveType<Native = T::Native>,

Reinterprets this array’s contents as a different data type without copying

This can be used to efficiently convert between primitive arrays with the same underlying representation

Note: this will not modify the underlying values, and therefore may change the semantic values of the array, e.g. 100 milliseconds in a TimestampNanosecondArray will become 100 seconds in a TimestampSecondArray.

For casts that preserve the semantic value, check out the [compute kernels]

compute kernels

let a = Int64Array::from_iter_values([1, 2, 3, 4]);
let b: TimestampNanosecondArray = a.reinterpret_cast();

pub fn unary<F, O>(&self, op: F) -> PrimitiveArray<O>where O: ArrowPrimitiveType, F: Fn(T::Native) -> O::Native,

Applies an unary and infallible function to a primitive array. This is the fastest way to perform an operation on a primitive array when the benefits of a vectorized operation outweigh the cost of branching nulls and non-nulls.

Implementation

This will apply the function for all values, including those on null slots. This implies that the operation must be infallible for any value of the corresponding type or this function may panic.

Example

let array = Int32Array::from(vec![Some(5), Some(7), None]);
let c = array.unary(|x| x * 2 + 1);
assert_eq!(c, Int32Array::from(vec![Some(11), Some(15), None]));

pub fn unary_mut<F>(self, op: F) -> Result<PrimitiveArray<T>, PrimitiveArray<T>>where F: Fn(T::Native) -> T::Native,

Applies an unary and infallible function to a mutable primitive array. Mutable primitive array means that the buffer is not shared with other arrays. As a result, this mutates the buffer directly without allocating new buffer.

Implementation

This will apply the function for all values, including those on null slots. This implies that the operation must be infallible for any value of the corresponding type or this function may panic.

Example

let array = Int32Array::from(vec![Some(5), Some(7), None]);
let c = array.unary_mut(|x| x * 2 + 1).unwrap();
assert_eq!(c, Int32Array::from(vec![Some(11), Some(15), None]));

pub fn try_unary<F, O, E>(&self, op: F) -> Result<PrimitiveArray<O>, E>where O: ArrowPrimitiveType, F: Fn(T::Native) -> Result<O::Native, E>,

Applies a unary and fallible function to all valid values in a primitive array

This is unlike Self::unary which will apply an infallible function to all rows regardless of validity, in many cases this will be significantly faster and should be preferred if op is infallible.

Note: LLVM is currently unable to effectively vectorize fallible operations

pub fn try_unary_mut<F, E>( self, op: F ) -> Result<Result<PrimitiveArray<T>, E>, PrimitiveArray<T>>where F: Fn(T::Native) -> Result<T::Native, E>,

Applies an unary and fallible function to all valid values in a mutable primitive array. Mutable primitive array means that the buffer is not shared with other arrays. As a result, this mutates the buffer directly without allocating new buffer.

This is unlike Self::unary_mut which will apply an infallible function to all rows regardless of validity, in many cases this will be significantly faster and should be preferred if op is infallible.

This returns an Err when the input array is shared buffer with other array. In the case, returned Err wraps input array. If the function encounters an error during applying on values. In the case, this returns an Err within an Ok which wraps the actual error.

Note: LLVM is currently unable to effectively vectorize fallible operations

pub fn unary_opt<F, O>(&self, op: F) -> PrimitiveArray<O>where O: ArrowPrimitiveType, F: Fn(T::Native) -> Option<O::Native>,

Applies a unary and nullable function to all valid values in a primitive array

This is unlike Self::unary which will apply an infallible function to all rows regardless of validity, in many cases this will be significantly faster and should be preferred if op is infallible.

Note: LLVM is currently unable to effectively vectorize fallible operations

pub fn into_builder(self) -> Result<PrimitiveBuilder<T>, Self>

Returns PrimitiveBuilder of this primitive array for mutating its values if the underlying data buffer is not shared by others.

impl<T: ArrowTemporalType> PrimitiveArray<T>where i64: From<T::Native>,

pub fn value_as_datetime(&self, i: usize) -> Option<NaiveDateTime>

Returns value as a chrono NaiveDateTime, handling time resolution

If a data type cannot be converted to NaiveDateTime, a None is returned. A valid value is expected, thus the user should first check for validity.

pub fn value_as_datetime_with_tz( &self, i: usize, tz: Tz ) -> Option<DateTime<Tz>>

Returns value as a chrono NaiveDateTime, handling time resolution with the provided tz

functionally it is same as value_as_datetime, however it adds the passed tz to the to-be-returned NaiveDateTime

pub fn value_as_date(&self, i: usize) -> Option<NaiveDate>

Returns value as a chrono NaiveDate by using Self::datetime()

If a data type cannot be converted to NaiveDate, a None is returned

pub fn value_as_time(&self, i: usize) -> Option<NaiveTime>

Returns a value as a chrono NaiveTime

Date32 and Date64 return UTC midnight as they do not have time resolution

pub fn value_as_duration(&self, i: usize) -> Option<Duration>

Returns a value as a chrono Duration

If a data type cannot be converted to Duration, a None is returned

impl<'a, T: ArrowPrimitiveType> PrimitiveArray<T>

pub fn iter(&'a self) -> PrimitiveIter<'a, T>

constructs a new iterator

impl<T: ArrowPrimitiveType> PrimitiveArray<T>

pub unsafe fn from_trusted_len_iter<I, P>(iter: I) -> Selfwhere P: Borrow<Option<<T as ArrowPrimitiveType>::Native>>, I: IntoIterator<Item = P>,

Creates a PrimitiveArray from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

impl<T: ArrowTimestampType> PrimitiveArray<T>

pub fn from_vec(data: Vec<i64>, timezone: Option<String>) -> Selfwhere Self: From<Vec<i64>>,

👎Deprecated: Use with_timezone_opt instead

Construct a timestamp array from a vec of i64 values and an optional timezone

pub fn from_opt_vec(data: Vec<Option<i64>>, timezone: Option<String>) -> Selfwhere Self: From<Vec<Option<i64>>>,

👎Deprecated: Use with_timezone_opt instead

Construct a timestamp array from a vec of Option<i64> values and an optional timezone

pub fn timezone(&self) -> Option<&str>

Returns the timezone of this array if any

pub fn with_timezone(self, timezone: impl Into<Arc<str>>) -> Self

Construct a timestamp array with new timezone

pub fn with_timezone_utc(self) -> Self

Construct a timestamp array with UTC

pub fn with_timezone_opt<S: Into<Arc<str>>>(self, timezone: Option<S>) -> Self

Construct a timestamp array with an optional timezone

impl<T: DecimalType + ArrowPrimitiveType> PrimitiveArray<T>

pub fn with_precision_and_scale( self, precision: u8, scale: i8 ) -> Result<Self, ArrowError>

Returns a Decimal array with the same data as self, with the specified precision and scale.

See validate_decimal_precision_and_scale

pub fn validate_decimal_precision( &self, precision: u8 ) -> Result<(), ArrowError>

Validates values in this array can be properly interpreted with the specified precision.

pub fn null_if_overflow_precision(&self, precision: u8) -> Self

Validates the Decimal Array, if the value of slot is overflow for the specified precision, and will be casted to Null

pub fn value_as_string(&self, row: usize) -> String

Returns Self::value formatted as a string

pub fn precision(&self) -> u8

Returns the decimal precision of this array

pub fn scale(&self) -> i8

Returns the decimal scale of this array