Type Alias arrow_array::array::Int8DictionaryArray

pub type Int8DictionaryArray = DictionaryArray<Int8Type>;

A DictionaryArray indexed by i8

Example: Using collect

let array: Int8DictionaryArray = vec!["a", "a", "b", "c"].into_iter().collect();
let values: Arc<dyn Array> = Arc::new(StringArray::from(vec!["a", "b", "c"]));
assert_eq!(array.keys(), &Int8Array::from(vec![0, 0, 1, 2]));
assert_eq!(array.values(), &values);

See DictionaryArray for more information and examples

Aliased Type

struct Int8DictionaryArray {
    data_type: DataType,
    keys: PrimitiveArray<Int8Type>,
    values: Arc<dyn Array, Global>,
    is_ordered: bool,
}

Fields
data_type: DataTypekeys: PrimitiveArray<Int8Type>

The keys of this dictionary. These are constructed from the buffer and null bitmap of data. Also, note that these do not correspond to the true values of this array. Rather, they map to the real values.

values: Arc<dyn Array, Global>

Array of dictionary values (can by any DataType).

is_ordered: bool

Values are ordered.

Implementations

impl<K: ArrowDictionaryKeyType> DictionaryArray<K>

pub fn new(keys: PrimitiveArray<K>, values: ArrayRef) -> Self

Attempt to create a new DictionaryArray with a specified keys (indexes into the dictionary) and values (dictionary) array.

Panics

Panics if Self::try_new returns an error

pub fn try_new( keys: PrimitiveArray<K>, values: ArrayRef ) -> Result<Self, ArrowError>

Attempt to create a new DictionaryArray with a specified keys (indexes into the dictionary) and values (dictionary) array.

Errors

Returns an error if any keys[i] >= values.len() || keys[i] < 0

pub unsafe fn new_unchecked(keys: PrimitiveArray<K>, values: ArrayRef) -> Self

Create a new DictionaryArray without performing validation

Safety

Safe provided Self::try_new would not return an error

pub fn into_parts(self) -> (PrimitiveArray<K>, ArrayRef)

Deconstruct this array into its constituent parts

pub fn keys(&self) -> &PrimitiveArray<K>

Return an array view of the keys of this dictionary as a PrimitiveArray.

pub fn lookup_key(&self, value: &str) -> Option<K::Native>

If value is present in values (aka the dictionary), returns the corresponding key (index into the values array). Otherwise returns None.

Panics if values is not a StringArray.

pub fn values(&self) -> &ArrayRef

Returns a reference to the dictionary values array

pub fn value_type(&self) -> DataType

Returns a clone of the value type of this list.

pub fn len(&self) -> usize

The length of the dictionary is the length of the keys array.

pub fn is_empty(&self) -> bool

Whether this dictionary is empty

pub fn is_ordered(&self) -> bool

Currently exists for compatibility purposes with Arrow IPC.

pub fn keys_iter(&self) -> impl Iterator<Item = Option<usize>> + '_

Return an iterator over the keys (indexes into the dictionary)

pub fn key(&self, i: usize) -> Option<usize>

Return the value of keys (the dictionary key) at index i, cast to usize, None if the value at i is NULL.

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a zero-copy slice of this array with the indicated offset and length.

pub fn downcast_dict<V: 'static>( &self ) -> Option<TypedDictionaryArray<'_, K, V>>

Downcast this dictionary to a TypedDictionaryArray

use arrow_array::{Array, ArrayAccessor, DictionaryArray, StringArray, types::Int32Type};

let orig = [Some("a"), Some("b"), None];
let dictionary = DictionaryArray::<Int32Type>::from_iter(orig);
let typed = dictionary.downcast_dict::<StringArray>().unwrap();
assert_eq!(typed.value(0), "a");
assert_eq!(typed.value(1), "b");
assert!(typed.is_null(2));

pub fn with_values(&self, values: ArrayRef) -> Self

Returns a new dictionary with the same keys as the current instance but with a different set of dictionary values

This can be used to perform an operation on the values of a dictionary

Panics

Panics if values has a length less than the current values

// Construct a Dict(Int32, Int8)
let mut builder = PrimitiveDictionaryBuilder::<Int32Type, Int8Type>::with_capacity(2, 200);
for i in 0..100 {
    builder.append(i % 2).unwrap();
}

let dictionary = builder.finish();

// Perform a widening cast of dictionary values
let typed_dictionary = dictionary.downcast_dict::<Int8Array>().unwrap();
let values: Int64Array = typed_dictionary.values().unary(|x| x as i64);

// Create a Dict(Int32,
let new = dictionary.with_values(Arc::new(values));

// Verify values are as expected
let new_typed = new.downcast_dict::<Int64Array>().unwrap();
for i in 0..100 {
    assert_eq!(new_typed.value(i), (i % 2) as i64)
}

pub fn into_primitive_dict_builder<V>( self ) -> Result<PrimitiveDictionaryBuilder<K, V>, Self>where V: ArrowPrimitiveType,

Returns PrimitiveDictionaryBuilder of this dictionary array for mutating its keys and values if the underlying data buffer is not shared by others.

pub fn unary_mut<F, V>( self, op: F ) -> Result<DictionaryArray<K>, DictionaryArray<K>>where V: ArrowPrimitiveType, F: Fn(V::Native) -> V::Native,

Applies an unary and infallible function to a mutable dictionary array. Mutable dictionary array means that the buffers are not shared with other arrays. As a result, this mutates the buffers directly without allocating new buffers.

Implementation

This will apply the function for all dictionary values, including those on null slots. This implies that the operation must be infallible for any value of the corresponding type or this function may panic.

Example

let values = Int32Array::from(vec![Some(10), Some(20), None]);
let keys = Int8Array::from_iter_values([0, 0, 1, 2]);
let dictionary = DictionaryArray::<Int8Type>::try_new(keys, Arc::new(values)).unwrap();
let c = dictionary.unary_mut::<_, Int32Type>(|x| x + 1).unwrap();
let typed = c.downcast_dict::<Int32Array>().unwrap();
assert_eq!(typed.value(0), 11);
assert_eq!(typed.value(1), 11);
assert_eq!(typed.value(2), 21);

pub fn occupancy(&self) -> BooleanBuffer

Computes an occupancy mask for this dictionary’s values

For each value in Self::values the corresponding bit will be set in the returned mask if it is referenced by a key in this DictionaryArray