Type Alias arrow_array::array::LargeBinaryArray

pub type LargeBinaryArray = GenericBinaryArray<i64>;

A GenericBinaryArray of [u8] using i64 offsets

Examples

Create a LargeBinaryArray from a vector of byte slices.

use arrow_array::{Array, LargeBinaryArray};
let values: Vec<&[u8]> =
    vec![b"one", b"two", b"", b"three"];
let array = LargeBinaryArray::from_vec(values);
assert_eq!(4, array.len());
assert_eq!(b"one", array.value(0));
assert_eq!(b"two", array.value(1));
assert_eq!(b"", array.value(2));
assert_eq!(b"three", array.value(3));

Create a LargeBinaryArray from a vector of Optional (null) byte slices.

use arrow_array::{Array, LargeBinaryArray};
let values: Vec<Option<&[u8]>> =
    vec![Some(b"one"), Some(b"two"), None, Some(b""), Some(b"three")];
let array = LargeBinaryArray::from_opt_vec(values);
assert_eq!(5, array.len());
assert_eq!(b"one", array.value(0));
assert_eq!(b"two", array.value(1));
assert_eq!(b"", array.value(3));
assert_eq!(b"three", array.value(4));
assert!(!array.is_null(0));
assert!(!array.is_null(1));
assert!(array.is_null(2));
assert!(!array.is_null(3));
assert!(!array.is_null(4));

See GenericByteArray for more information and examples

Aliased Type

struct LargeBinaryArray {
    data_type: DataType,
    value_offsets: OffsetBuffer<i64>,
    value_data: Buffer,
    nulls: Option<NullBuffer>,
}

Fields
data_type: DataTypevalue_offsets: OffsetBuffer<i64>value_data: Buffernulls: Option<NullBuffer>
Implementations

impl<OffsetSize: OffsetSizeTrait> GenericByteArray<GenericBinaryType<OffsetSize>>

pub const fn get_data_type() -> DataType

๐Deprecated: please use Self::DATA_TYPE instead

Get the data type of the array.

pub fn from_vec(v: Vec<&[u8]>) -> Self

Creates a GenericBinaryArray from a vector of byte slices

See also Self::from_iter_values

pub fn from_opt_vec(v: Vec<Option<&[u8]>>) -> Self

Creates a GenericBinaryArray from a vector of Optional (null) byte slices

fn from_list(v: GenericListArray<OffsetSize>) -> Self

pub fn take_iter<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&[u8]>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&[u8]>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the indexes in the iterator are less than the array.len()

impl<T: ByteArrayType> GenericByteArray<T>

pub const DATA_TYPE: DataType = T::DATA_TYPE

Data type of the array.

pub fn new( offsets: OffsetBuffer<T::Offset>, values: Buffer, nulls: Option<NullBuffer> ) -> Self

Create a new GenericByteArray from the provided parts, panicking on failure

Panics

Panics if GenericByteArray::try_new returns an error

pub fn try_new( offsets: OffsetBuffer<T::Offset>, values: Buffer, nulls: Option<NullBuffer> ) -> Result<Self, ArrowError>

Create a new GenericByteArray from the provided parts, returning an error on failure

Errors

offsets.len() - 1 != nulls.len()

Any consecutive pair of offsets does not denote a valid slice of values

pub unsafe fn new_unchecked( offsets: OffsetBuffer<T::Offset>, values: Buffer, nulls: Option<NullBuffer> ) -> Self

Create a new GenericByteArray from the provided parts, without validation

Safety

Safe if Self::try_new would not error

pub fn new_null(len: usize) -> Self

Create a new GenericByteArray of length len where all values are null

pub fn new_scalar(value: impl AsRef<T::Native>) -> Scalar<Self>

Create a new Scalar from v

pub fn from_iter_values<Ptr, I>(iter: I) -> Selfwhere Ptr: AsRef<T::Native>, I: IntoIterator<Item = Ptr>,

Creates a GenericByteArray based on an iterator of values without nulls

pub fn into_parts(self) -> (OffsetBuffer<T::Offset>, Buffer, Option<NullBuffer>)

Deconstruct this array into its constituent parts

pub fn value_length(&self, i: usize) -> T::Offset

Returns the length for value at index i.

Panics

Panics if index i is out of bounds.

pub fn offsets(&self) -> &OffsetBuffer<T::Offset>

Returns a reference to the offsets of this array

Unlike Self::value_offsets this returns the [OffsetBuffer] allowing for zero-copy cloning

pub fn values(&self) -> &Buffer

Returns the values of this array

Unlike Self::value_data this returns the [Buffer] allowing for zero-copy cloning

pub fn value_data(&self) -> &[u8]

Returns the raw value data

pub fn is_ascii(&self) -> bool

Returns true if all data within this array is ASCII

pub fn value_offsets(&self) -> &[T::Offset]

Returns the offset values in the offsets buffer

pub unsafe fn value_unchecked(&self, i: usize) -> &T::Native

Returns the element at index i

Safety

Caller is responsible for ensuring that the index is within the bounds of the array

pub fn value(&self, i: usize) -> &T::Native

Returns the element at index i

Panics

Panics if index i is out of bounds.

pub fn iter(&self) -> ArrayIter<&Self>

constructs a new iterator

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a zero-copy slice of this array with the indicated offset and length.

pub fn into_builder(self) -> Result<GenericByteBuilder<T>, Self>

Returns GenericByteBuilder of this byte array for mutating its values if the underlying offset and data buffers are not shared by others.