arrow_array::array

Module union_array

Structs

UnionArray

An array of values of varying types

Enums

Mask 🔒

SparseStrategy 🔒

How to compute the logical nulls of a sparse union. All strategies return the same result. Those starting with Mask perform bitwise masking for each chunk of 64 values, including computing expensive selection masks of fields: which fields masks must be computed is the difference between them

Functions

selection_mask 🔒

without_nulls_selected 🔒

Returns a bitmask where bits indicate if any id from without_nulls_ids exist in type_ids_chunk.