Module arrow_array::builder

Defines builders for the various array types

Structs

BooleanBufferBuilder

A builder for creating a boolean Buffer

BooleanBuilder

Array builder for fixed-width primitive types

BufferBuilder

Builder for creating a Buffer object.

FixedSizeBinaryBuilder

A fixed size binary array builder

FixedSizeListBuilder

Array builder for FixedSizeListArray

GenericByteBuilder

Array builder for GenericByteArray

GenericByteDictionaryBuilder

Generic array builder for DictionaryArray that stores generic byte values. For example to map a set of byte indices to String values. Note that the use of a HashMap here will not scale to very large arrays or result in an ordered dictionary.

GenericByteRunBuilder

Array builder for RunArray for String and Binary types.

GenericListBuilder

Array builder for GenericListArray

MapBuilder

Creates a new MapBuilder

MapFieldNames

Contains details of the mapping

PrimitiveBuilder

Array builder for fixed-width primitive types

PrimitiveDictionaryBuilder

Array builder for DictionaryArray. For example to map a set of byte indices to f32 values. Note that the use of a HashMap here will not scale to very large arrays or result in an ordered dictionary.

PrimitiveRunBuilder

Array builder for RunArray that encodes primitive values.

StructBuilder

Array builder for Struct types.

UnionBuilder

Builder type for creating a new UnionArray.

Traits

ArrayBuilder

Trait for dealing with different array builders at runtime

Functions

make_builder

Returns a builder with capacity capacity that corresponds to the datatype DataType This function is useful to construct arrays from an arbitrary vectors with known/expected schema.

Type Definitions

BinaryBuilder

A binary array builder with i32 offsets

BinaryDictionaryBuilder

Array builder for DictionaryArray that stores binary. For example to map a set of byte indices to binary values. Note that the use of a HashMap here will not scale to very large arrays or result in an ordered dictionary.

BinaryRunBuilder

Array builder for RunArray that encodes binary values(BinaryType).

Date32BufferBuilder

Buffer builder for 32-bit date type.

Date32Builder

A 32-bit date array builder.

Date64BufferBuilder

Buffer builder for 64-bit date type.

Date64Builder

A 64-bit date array builder.

Decimal128Builder

A decimal 128 array builder

Decimal256Builder

A decimal 256 array builder

DurationMicrosecondBufferBuilder

Buffer builder for elaspsed time of microseconds unit.

DurationMicrosecondBuilder

An elapsed time in microseconds array builder.

DurationMillisecondBufferBuilder

Buffer builder for elaspsed time of milliseconds unit.

DurationMillisecondBuilder

An elapsed time in milliseconds array builder.

DurationNanosecondBufferBuilder

Buffer builder for elaspsed time of nanoseconds unit.

DurationNanosecondBuilder

An elapsed time in nanoseconds array builder.

DurationSecondBufferBuilder

Buffer builder for elaspsed time of second unit.

DurationSecondBuilder

An elapsed time in seconds array builder.

Float32BufferBuilder

Buffer builder for 32-bit floating point type.

Float32Builder

A 32-bit floating point array builder.

Float64BufferBuilder

Buffer builder for 64-bit floating point type.

Float64Builder

A 64-bit floating point array builder.

GenericBinaryBuilder

Array builder for GenericBinaryArray

GenericStringBuilder

Array builder for GenericStringArray

Int8BufferBuilder

Buffer builder for signed 8-bit integer type.

Int8Builder

A signed 8-bit integer array builder.

Int16BufferBuilder

Buffer builder for signed 16-bit integer type.

Int16Builder

A signed 16-bit integer array builder.

Int32BufferBuilder

Buffer builder for signed 32-bit integer type.

Int32Builder

A signed 32-bit integer array builder.

Int64BufferBuilder

Buffer builder for signed 64-bit integer type.

Int64Builder

A signed 64-bit integer array builder.

IntervalDayTimeBufferBuilder

Buffer builder for “calendar” interval in days and milliseconds.

IntervalDayTimeBuilder

A “calendar” interval in days and milliseconds array builder.

IntervalMonthDayNanoBufferBuilder

Buffer builder “calendar” interval in months, days, and nanoseconds.

IntervalMonthDayNanoBuilder

A “calendar” interval in months, days, and nanoseconds array builder.

IntervalYearMonthBufferBuilder

Buffer builder for “calendar” interval in months.

IntervalYearMonthBuilder

A “calendar” interval in months array builder.

LargeBinaryBuilder

A binary array builder with i64 offsets

LargeBinaryDictionaryBuilder

Array builder for DictionaryArray that stores large binary. For example to map a set of byte indices to binary values. Note that the use of a HashMap here will not scale to very large arrays or result in an ordered dictionary.

LargeBinaryRunBuilder

Array builder for RunArray that encodes large binary values(LargeBinaryType). See documentation of BinaryRunBuilder for an example.

LargeListBuilder

A list array builder with i64 offsets

LargeStringBuilder

A string array builder with i64 offsets

LargeStringDictionaryBuilder

Array builder for DictionaryArray that stores large Strings. For example to map a set of byte indices to String values. Note that the use of a HashMap here will not scale to very large arrays or result in an ordered dictionary.

LargeStringRunBuilder

Array builder for RunArray that encodes large strings (LargeUtf8Type). See StringRunBuilder for an example.

ListBuilder

A list array builder with i32 offsets

StringBuilder

A string array builder with i32 offsets

StringDictionaryBuilder

Array builder for DictionaryArray that stores Strings. For example to map a set of byte indices to String values. Note that the use of a HashMap here will not scale to very large arrays or result in an ordered dictionary.

StringRunBuilder

Array builder for RunArray that encodes strings (Utf8Type).

Time32MillisecondBufferBuilder

Buffer builder for 32-bit elaspsed time since midnight of millisecond unit.

Time32MillisecondBuilder

A 32-bit elaspsed time in milliseconds array builder.

Time32SecondBufferBuilder

Buffer builder for 32-bit elaspsed time since midnight of second unit.

Time32SecondBuilder

A 32-bit elaspsed time in seconds array builder.

Time64MicrosecondBufferBuilder

Buffer builder for 64-bit elaspsed time since midnight of microsecond unit.

Time64MicrosecondBuilder

A 64-bit elaspsed time in microseconds array builder.

Time64NanosecondBufferBuilder

Buffer builder for 64-bit elaspsed time since midnight of nanosecond unit.

Time64NanosecondBuilder

A 64-bit elaspsed time in nanoseconds array builder.

TimestampMicrosecondBufferBuilder

Buffer builder for timestamp type of microsecond unit.

TimestampMicrosecondBuilder

A timestamp microsecond array builder.

TimestampMillisecondBufferBuilder

Buffer builder for timestamp type of millisecond unit.

TimestampMillisecondBuilder

A timestamp millisecond array builder.

TimestampNanosecondBufferBuilder

Buffer builder for timestamp type of nanosecond unit.

TimestampNanosecondBuilder

A timestamp nanosecond array builder.

TimestampSecondBufferBuilder

Buffer builder for timestamp type of second unit.

TimestampSecondBuilder

A timestamp second array builder.

UInt8BufferBuilder

Buffer builder for usigned 8-bit integer type.

UInt8Builder

An usigned 8-bit integer array builder.

UInt16BufferBuilder

Buffer builder for usigned 16-bit integer type.

UInt16Builder

An usigned 16-bit integer array builder.

UInt32BufferBuilder

Buffer builder for usigned 32-bit integer type.

UInt32Builder

An usigned 32-bit integer array builder.

UInt64BufferBuilder

Buffer builder for usigned 64-bit integer type.

UInt64Builder

An usigned 64-bit integer array builder.