arrow_array::builder

Function make_builder

pub fn make_builder(
    datatype: &DataType,
    capacity: usize,
) -> Box<dyn ArrayBuilder>

Returns a builder with capacity for capacity elements of datatype DataType.

This function is useful to construct arrays from an arbitrary vectors with known/expected schema.

See comments on StructBuilder for retrieving collection builders built by make_builder.