Struct arrow_array::builder::generic_byte_run_builder::GenericByteRunBuilder

pub struct GenericByteRunBuilder<R, V>where
    R: ArrowPrimitiveType,
    V: ByteArrayType,{
    run_ends_builder: PrimitiveBuilder<R>,
    values_builder: GenericByteBuilder<V>,
    current_value: Vec<u8>,
    has_current_value: bool,
    current_run_end_index: usize,
    prev_run_end_index: usize,
}

Builder for RunArray of GenericByteArray

Example:

let mut builder =
GenericByteRunBuilder::<Int16Type, BinaryType>::new();
builder.extend([Some(b"abc"), Some(b"abc"), None, Some(b"def")].into_iter());
builder.append_value(b"def");
builder.append_null();
let array = builder.finish();

assert_eq!(array.run_ends().values(), &[2, 3, 5, 6]);

let av = array.values();

assert!(!av.is_null(0));
assert!(av.is_null(1));
assert!(!av.is_null(2));
assert!(av.is_null(3));

// Values are polymorphic and so require a downcast.
let ava: &BinaryArray = av.as_binary();

assert_eq!(ava.value(0), b"abc");
assert_eq!(ava.value(2), b"def");

Fields
run_ends_builder: PrimitiveBuilder<R>values_builder: GenericByteBuilder<V>current_value: Vec<u8>has_current_value: boolcurrent_run_end_index: usizeprev_run_end_index: usize
Implementations

impl<R, V> GenericByteRunBuilder<R, V>where R: ArrowPrimitiveType, V: ByteArrayType,

pub fn new() -> Self

Creates a new GenericByteRunBuilder

pub fn with_capacity(capacity: usize, data_capacity: usize) -> Self

Creates a new GenericByteRunBuilder with the provided capacity

capacity: the expected number of run-end encoded values. data_capacity: the expected number of bytes of run end encoded values

impl<R, V> GenericByteRunBuilder<R, V>where R: RunEndIndexType, V: ByteArrayType,

pub fn append_option(&mut self, input_value: Option<impl AsRef<V::Native>>)

Appends optional value to the logical array encoded by the RunArray.

pub fn append_value(&mut self, input_value: impl AsRef<V::Native>)

Appends value to the logical array encoded by the RunArray.

pub fn append_null(&mut self)

Appends null to the logical array encoded by the RunArray.

pub fn finish(&mut self) -> RunArray<R>

Creates the RunArray and resets the builder. Panics if RunArray cannot be built.

pub fn finish_cloned(&self) -> RunArray<R>

Creates the RunArray and without resetting the builder. Panics if RunArray cannot be built.

fn append_run_end(&mut self)

fn append_run_end_with_builders( &self, run_ends_builder: &mut PrimitiveBuilder<R>, values_builder: &mut GenericByteBuilder<V> )

fn run_end_index_as_native(&self) -> R::Native