Type Alias arrow_array::builder::generic_byte_run_builder::LargeStringRunBuilder

pub type LargeStringRunBuilder<K> = GenericByteRunBuilder<K, LargeUtf8Type>;

Builder for RunArray of LargeStringArray

Aliased Type

struct LargeStringRunBuilder<K> {
    run_ends_builder: PrimitiveBuilder<K>,
    values_builder: GenericByteBuilder<GenericStringType<i64>>,
    current_value: Vec<u8, Global>,
    has_current_value: bool,
    current_run_end_index: usize,
    prev_run_end_index: usize,
}

Fields
run_ends_builder: PrimitiveBuilder<K>values_builder: GenericByteBuilder<GenericStringType<i64>>current_value: Vec<u8, Global>has_current_value: boolcurrent_run_end_index: usizeprev_run_end_index: usize
Implementations

impl<R, V> GenericByteRunBuilder<R, V>where R: ArrowPrimitiveType, V: ByteArrayType,

pub fn new() -> Self

Creates a new GenericByteRunBuilder

pub fn with_capacity(capacity: usize, data_capacity: usize) -> Self

Creates a new GenericByteRunBuilder with the provided capacity

capacity: the expected number of run-end encoded values. data_capacity: the expected number of bytes of run end encoded values

impl<R, V> GenericByteRunBuilder<R, V>where R: RunEndIndexType, V: ByteArrayType,

pub fn append_option(&mut self, input_value: Option<impl AsRef<V::Native>>)

Appends optional value to the logical array encoded by the RunArray.

pub fn append_value(&mut self, input_value: impl AsRef<V::Native>)

Appends value to the logical array encoded by the RunArray.

pub fn append_null(&mut self)

Appends null to the logical array encoded by the RunArray.

pub fn finish(&mut self) -> RunArray<R>

Creates the RunArray and resets the builder. Panics if RunArray cannot be built.

pub fn finish_cloned(&self) -> RunArray<R>

Creates the RunArray and without resetting the builder. Panics if RunArray cannot be built.

fn append_run_end(&mut self)

fn append_run_end_with_builders( &self, run_ends_builder: &mut PrimitiveBuilder<R>, values_builder: &mut GenericByteBuilder<V> )

fn run_end_index_as_native(&self) -> R::Native