arrow_array::builder

Module generic_bytes_builder

Structs

GenericByteBuilder

Builder for GenericByteArray

Constants

AVERAGE_BINARY_LENGTH 🔒

A byte size value representing the number of bytes to allocate per binary in GenericBinaryBuilder

AVERAGE_STRING_LENGTH 🔒

A byte size value representing the number of bytes to allocate per string in GenericStringBuilder

Traits

BinaryLikeArrayBuilder

Trait for binary-like array builders

StringLikeArrayBuilder

Trait for string-like array builders

Type Aliases

GenericBinaryBuilder

Array builder for GenericBinaryArray

GenericStringBuilder

Array builder for GenericStringArray