Module arrow_array::builder::generic_bytes_view_builder

Structs

GenericByteViewBuilder

A builder for GenericByteViewArray

Enums

BlockSizeGrowthStrategy 🔒

Constants

MAX_BLOCK_SIZE 🔒

STARTING_BLOCK_SIZE 🔒

Functions

make_inlined_view 🔒

Creates a view from a fixed length input (the compiler can generate specialized code for this)

make_view

Create a view based on the given data, block id and offset Note that the code below is carefully examined with x86_64 assembly code: https://godbolt.org/z/685YPsd5G The goal is to avoid calling into ptr::copy_non_interleave, which makes function call (i.e., not inlined), which slows down things.

Type Aliases

BinaryViewBuilder

Array builder for BinaryViewArray

StringViewBuilder

Array builder for StringViewArray