arrow_array::builder

Struct PrimitiveRunBuilder

pub struct PrimitiveRunBuilder<R, V>
where
    R: RunEndIndexType,
    V: ArrowPrimitiveType,
{
    run_ends_builder: PrimitiveBuilder<R>,
    values_builder: PrimitiveBuilder<V>,
    current_value: Option<V::Native>,
    current_run_end_index: usize,
    prev_run_end_index: usize,
}

Builder for RunArray of PrimitiveArray

Example:

let mut builder =
PrimitiveRunBuilder::<Int16Type, UInt32Type>::new();
builder.append_value(1234);
builder.append_value(1234);
builder.append_value(1234);
builder.append_null();
builder.append_value(5678);
builder.append_value(5678);
let array = builder.finish();

assert_eq!(array.run_ends().values(), &[3, 4, 6]);

let av = array.values();

assert!(!av.is_null(0));
assert!(av.is_null(1));
assert!(!av.is_null(2));

// Values are polymorphic and so require a downcast.
let ava: &UInt32Array = av.as_primitive::<UInt32Type>();

assert_eq!(ava, &UInt32Array::from(vec![Some(1234), None, Some(5678)]));

Fields
run_ends_builder: PrimitiveBuilder<R>values_builder: PrimitiveBuilder<V>current_value: Option<V::Native>current_run_end_index: usizeprev_run_end_index: usize
Implementations

impl<R, V> PrimitiveRunBuilder<R, V>
where R: RunEndIndexType, V: ArrowPrimitiveType,

pub fn new() -> Self

Creates a new PrimitiveRunBuilder

pub fn with_capacity(capacity: usize) -> Self

Creates a new PrimitiveRunBuilder with the provided capacity

capacity: the expected number of run-end encoded values.

impl<R, V> PrimitiveRunBuilder<R, V>
where R: RunEndIndexType, V: ArrowPrimitiveType,

pub fn append_option(&mut self, value: Option<V::Native>)

Appends optional value to the logical array encoded by the RunArray.

pub fn append_value(&mut self, value: V::Native)

Appends value to the logical array encoded by the run-ends array.

pub fn append_null(&mut self)

Appends null to the logical array encoded by the run-ends array.

pub fn finish(&mut self) -> RunArray<R>

Creates the RunArray and resets the builder. Panics if RunArray cannot be built.

pub fn finish_cloned(&self) -> RunArray<R>

Creates the RunArray and without resetting the builder. Panics if RunArray cannot be built.

fn append_run_end(&mut self)

fn append_run_end_with_builders( &self, run_ends_builder: &mut PrimitiveBuilder<R>, values_builder: &mut PrimitiveBuilder<V>, )

fn run_end_index_as_native(&self) -> R::Native