arrow_array::builder::union_builder
Struct FieldData

struct FieldData {
    type_id: i8,
    data_type: DataType,
    values_buffer: Box<dyn FieldDataValues>,
    slots: usize,
    null_buffer_builder: NullBufferBuilder,
}

FieldData is a helper struct to track the state of the fields in the UnionBuilder.

Fields
type_id: i8

The type id for this field

data_type: DataType

The Arrow data type represented in the values_buffer, which is untyped

values_buffer: Box<dyn FieldDataValues>

A buffer containing the values for this field in raw bytes

slots: usize

The number of array slots represented by the buffer

null_buffer_builder: NullBufferBuilder

A builder for the null bitmap

Implementations

impl FieldData

fn new<T: ArrowPrimitiveType>( type_id: i8, data_type: DataType, capacity: usize, ) -> Self

Creates a new FieldData.

fn append_value<T: ArrowPrimitiveType>(&mut self, v: T::Native)

Appends a single value to this FieldData’s values_buffer.

fn append_null(&mut self)

Appends a null to this FieldData.