Module arrow_array::cast

Defines helper functions for downcasting dyn Array to concrete types

Functions

as_boolean_array

Force downcast of an Array, such as an ArrayRef to BooleanArray, panic’ing on failure.

as_decimal_array

Force downcast of an Array, such as an ArrayRef to Decimal128Array, panic’ing on failure.

as_dictionary_array

Force downcast of an Array, such as an ArrayRef to DictionaryArray<T>, panic’ing on failure.

as_generic_binary_array

Force downcast of an Array, such as an ArrayRef to GenericBinaryArray<S>, panic’ing on failure.

as_generic_list_array

Force downcast of an Array, such as an ArrayRef to GenericListArray<T>, panic’ing on failure.

as_large_list_array

Force downcast of an Array, such as an ArrayRef to LargeListArray, panic’ing on failure.

as_largestring_array

Force downcast of an Array, such as an ArrayRef to LargeStringArray, panic’ing on failure.

as_list_array

Force downcast of an Array, such as an ArrayRef to ListArray, panic’ing on failure.

as_map_array

Force downcast of an Array, such as an ArrayRef to MapArray, panic’ing on failure.

as_null_array

Force downcast of an Array, such as an ArrayRef to NullArray, panic’ing on failure.

as_primitive_array

Force downcast of an Array, such as an ArrayRef, to PrimitiveArray<T>, panic’ing on failure.

as_string_array

Force downcast of an Array, such as an ArrayRef to StringArray, panic’ing on failure.

as_struct_array

Force downcast of an Array, such as an ArrayRef to StructArray, panic’ing on failure.

as_union_array

Force downcast of an Array, such as an ArrayRef to UnionArray, panic’ing on failure.

downcast_array

Downcasts a dyn Array to a concrete type