arrow_array

Module delta

Contains utility functions for shifting Date objects.

Functions

add_days_datetime 🔒

Add the given number of days to the given datetime.

add_months_datetime 🔒

Add the given number of months to the given datetime.

shift_months 🔒

Shift a date by the given number of months.

sub_days_datetime 🔒

Substract the given number of days to the given datetime.

sub_months_datetime 🔒

Substract the given number of months to the given datetime.