arrow_array

Module scalar

Structs

Scalar

A wrapper around a single value Array that implements Datum and indicates compute kernels should treat this array as a scalar value (a single value).

Traits

Datum

A possibly Scalar Array