Trait arrow_array::types::ArrowPrimitiveType

pub trait ArrowPrimitiveType: PrimitiveTypeSealed + 'static {
    type Native: ArrowNativeTypeOp;

    const DATA_TYPE: DataType;

    // Provided methods
    fn get_byte_width() -> usize { ... }
    fn default_value() -> Self::Native { ... }
}

Trait for primitive values, bridging the dynamic-typed nature of Arrow (via [DataType]) with the static-typed nature of rust types (ArrowNativeType) for all types that implement ArrowNativeType.

Required Associated Types

type Native: ArrowNativeTypeOp

Corresponding Rust native type for the primitive type.

Required Associated Constants

const DATA_TYPE: DataType

the corresponding Arrow data type of this primitive type.

Provided Methods

fn get_byte_width() -> usize

๐Deprecated: Use ArrowNativeType::get_byte_width

Returns the byte width of this primitive type.

fn default_value() -> Self::Native

Returns a default value of this primitive type.

This is useful for aggregate array ops like sum(), mean().

Object Safety

This trait is not object safe.