Struct arrow_array::array::GenericByteArray

pub struct GenericByteArray<T: ByteArrayType> { /* private fields */ }

Generic struct for variable-size byte arrays

See StringArray and LargeStringArray for storing utf8 encoded string data

See BinaryArray and LargeBinaryArray for storing arbitrary bytes

Implementations

impl<OffsetSize: OffsetSizeTrait> GenericByteArray<GenericBinaryType<OffsetSize>>

pub const fn get_data_type() -> DataType

👎Deprecated: please use Self::DATA_TYPE instead

Get the data type of the array.

pub fn from_vec(v: Vec<&[u8]>) -> Self

Creates a GenericBinaryArray from a vector of byte slices

See also Self::from_iter_values

pub fn from_opt_vec(v: Vec<Option<&[u8]>>) -> Self

Creates a GenericBinaryArray from a vector of Optional (null) byte slices

pub fn from_iter_values<Ptr, I>(iter: I) -> Selfwhere Ptr: AsRef<[u8]>, I: IntoIterator<Item = Ptr>,

Creates a GenericBinaryArray based on an iterator of values without nulls

pub fn take_iter<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&[u8]>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&[u8]>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the indexes in the iterator are less than the array.len()

impl<T: ByteArrayType> GenericByteArray<T>

pub const DATA_TYPE: DataType = T::DATA_TYPE

Data type of the array.

pub fn value_length(&self, i: usize) -> T::Offset

Returns the length for value at index i.

Panics

Panics if index i is out of bounds.

pub fn value_data(&self) -> &[u8]

Returns the raw value data

pub fn is_ascii(&self) -> bool

Returns true if all data within this array is ASCII

pub fn value_offsets(&self) -> &[T::Offset]

Returns the offset values in the offsets buffer

pub unsafe fn value_unchecked(&self, i: usize) -> &T::Native

Returns the element at index i

Safety

Caller is responsible for ensuring that the index is within the bounds of the array

pub fn value(&self, i: usize) -> &T::Native

Returns the element at index i

Panics

Panics if index i is out of bounds.

pub fn iter(&self) -> ArrayIter<&Self>

constructs a new iterator

pub fn into_builder(self) -> Result<GenericByteBuilder<T>, Self>

Returns GenericByteBuilder of this byte array for mutating its values if the underlying offset and data buffers are not shared by others.

impl<OffsetSize: OffsetSizeTrait> GenericByteArray<GenericStringType<OffsetSize>>

pub const fn get_data_type() -> DataType

👎Deprecated: please use Self::DATA_TYPE instead

Get the data type of the array.

pub fn num_chars(&self, i: usize) -> usize

Returns the number of Unicode Scalar Value in the string at index i.

Performance

This function has O(n) time complexity where n is the string length. If you can make sure that all chars in the string are in the range U+0x0000 ~ U+0x007F, please use the function value_length which has O(1) time complexity.

pub fn from_iter_values<Ptr, I>(iter: I) -> Selfwhere Ptr: AsRef<str>, I: IntoIterator<Item = Ptr>,

Creates a GenericStringArray based on an iterator of values without nulls

pub fn take_iter<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&str>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&str>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the indexes in the iterator are less than the array.len()