arrow_array::array
Type Alias GenericBinaryArray

pub type GenericBinaryArray<OffsetSize> = GenericByteArray<GenericBinaryType<OffsetSize>>;

A GenericBinaryArray for storing [u8]

Aliased Type

struct GenericBinaryArray<OffsetSize> {
    data_type: DataType,
    value_offsets: OffsetBuffer<<GenericBinaryType<OffsetSize> as ByteArrayType>::Offset>,
    value_data: Buffer,
    nulls: Option<NullBuffer>,
}

Fields
data_type: DataTypevalue_offsets: OffsetBuffer<<GenericBinaryType<OffsetSize> as ByteArrayType>::Offset>value_data: Buffernulls: Option<NullBuffer>
Implementations

impl<OffsetSize: OffsetSizeTrait> GenericBinaryArray<OffsetSize>

pub const fn get_data_type() -> DataType

๐Deprecated: please use Self::DATA_TYPE instead

Get the data type of the array.

pub fn from_vec(v: Vec<&[u8]>) -> Self

Creates a GenericBinaryArray from a vector of byte slices

See also Self::from_iter_values

pub fn from_opt_vec(v: Vec<Option<&[u8]>>) -> Self

Creates a GenericBinaryArray from a vector of Optional (null) byte slices

fn from_list(v: GenericListArray<OffsetSize>) -> Self

pub fn take_iter<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a, ) -> impl Iterator<Item = Option<&'a [u8]>>

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a, ) -> impl Iterator<Item = Option<&'a [u8]>>

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the indexes in the iterator are less than the array.len()