Type Definition arrow_array::array::GenericStringArray

pub type GenericStringArray<OffsetSize> = GenericByteArray<GenericStringType<OffsetSize>>;

Generic struct for [Large]StringArray

See StringArray and LargeStringArray for storing specific string data.

Implementations

impl<OffsetSize: OffsetSizeTrait> GenericStringArray<OffsetSize>

pub const fn get_data_type() -> DataType

👎Deprecated: please use Self::DATA_TYPE instead

Get the data type of the array.

pub fn num_chars(&self, i: usize) -> usize

Returns the number of Unicode Scalar Value in the string at index i.

Performance

This function has O(n) time complexity where n is the string length. If you can make sure that all chars in the string are in the range U+0x0000 ~ U+0x007F, please use the function value_length which has O(1) time complexity.

pub fn from_iter_values<Ptr, I>(iter: I) -> Selfwhere Ptr: AsRef<str>, I: IntoIterator<Item = Ptr>,

Creates a GenericStringArray based on an iterator of values without nulls

pub fn take_iter<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&str>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>( &'a self, indexes: impl Iterator<Item = Option<usize>> + 'a ) -> impl Iterator<Item = Option<&str>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the indexes in the iterator are less than the array.len()