arrow_array::array
Type Alias Int8DictionaryArray

pub type Int8DictionaryArray = DictionaryArray<Int8Type>;

A DictionaryArray indexed by i8

Example: Using collect

let array: Int8DictionaryArray = vec!["a", "a", "b", "c"].into_iter().collect();
let values: Arc<dyn Array> = Arc::new(StringArray::from(vec!["a", "b", "c"]));
assert_eq!(array.keys(), &Int8Array::from(vec![0, 0, 1, 2]));
assert_eq!(array.values(), &values);

See DictionaryArray for more information and examples

Aliased Type

struct Int8DictionaryArray {
    data_type: DataType,
    keys: PrimitiveArray<Int8Type>,
    values: Arc<dyn Array>,
    is_ordered: bool,
}

Fields
data_type: DataTypekeys: PrimitiveArray<Int8Type>

The keys of this dictionary. These are constructed from the buffer and null bitmap of data. Also, note that these do not correspond to the true values of this array. Rather, they map to the real values.

values: Arc<dyn Array>

Array of dictionary values (can by any DataType).

is_ordered: bool

Values are ordered.