arrow_array::builder

Module generic_bytes_view_builder

Structs

GenericByteViewBuilder

A builder for GenericByteViewArray

Enums

BlockSizeGrowthStrategy 🔒

Constants

MAX_BLOCK_SIZE 🔒

STARTING_BLOCK_SIZE 🔒

Functions

make_inlined_view 🔒

Creates a view from a fixed length input (the compiler can generate specialized code for this)

make_view

Create a view based on the given data, block id and offset.

Type Aliases

BinaryViewBuilder

Array builder for BinaryViewArray

StringViewBuilder

Array builder for StringViewArray