Struct arrow_array::builder::GenericListBuilder

pub struct GenericListBuilder<OffsetSize: OffsetSizeTrait, T: ArrayBuilder> { /* private fields */ }

Array builder for GenericListArray

Implementations

impl<OffsetSize: OffsetSizeTrait, T: ArrayBuilder> GenericListBuilder<OffsetSize, T>

pub fn new(values_builder: T) -> Self

Creates a new GenericListBuilder from a given values array builder

pub fn with_capacity(values_builder: T, capacity: usize) -> Self

Creates a new GenericListBuilder from a given values array builder capacity is the number of items to pre-allocate space for in this builder

impl<OffsetSize: OffsetSizeTrait, T> GenericListBuilder<OffsetSize, T>where T: 'static + ArrayBuilder,

pub fn values(&mut self) -> &mut T

Returns the child array builder as a mutable reference.

This mutable reference can be used to append values into the child array builder, but you must call append to delimit each distinct list value.

pub fn values_ref(&self) -> &T

Returns the child array builder as an immutable reference

pub fn append(&mut self, is_valid: bool)

Finish the current variable-length list array slot

Panics

Panics if the length of Self::values exceeds OffsetSize::MAX

pub fn finish(&mut self) -> GenericListArray<OffsetSize>

Builds the GenericListArray and reset this builder.

pub fn finish_cloned(&self) -> GenericListArray<OffsetSize>

Builds the GenericListArray without resetting the builder.

pub fn offsets_slice(&self) -> &[OffsetSize]

Returns the current offsets buffer as a slice