arrow_array

Module cast

Defines helper functions for downcasting dyn Array to concrete types

Modules

private 🔒

Macros

array_downcast_fn 🔒

Traits

AsArray

An extension trait for dyn Array that provides ergonomic downcasting

Functions

as_boolean_array

Force downcast of an Array, such as an ArrayRef to BooleanArray, panicking on failure.

as_dictionary_array

Force downcast of an Array, such as an ArrayRef to DictionaryArray<T>, panic’ing on failure.

as_fixed_size_list_array

Force downcast of an Array, such as an ArrayRef to FixedSizeListArray, panicking on failure.

as_generic_binary_array

Force downcast of an Array, such as an ArrayRef to GenericBinaryArray<S>, panicking on failure.

as_generic_list_array

Force downcast of an Array, such as an ArrayRef to GenericListArray<T>, panicking on failure.

as_large_list_array

Force downcast of an Array, such as an ArrayRef to LargeListArray, panicking on failure.

as_largestring_array

Force downcast of an Array, such as an ArrayRef to LargeStringArray, panicking on failure.

as_list_array

Force downcast of an Array, such as an ArrayRef to ListArray, panicking on failure.

as_map_array

Force downcast of an Array, such as an ArrayRef to MapArray, panicking on failure.

as_null_array

Force downcast of an Array, such as an ArrayRef to NullArray, panicking on failure.

as_primitive_array

Force downcast of an Array, such as an ArrayRef, to PrimitiveArray<T>, panic’ing on failure.

as_run_array

Force downcast of an Array, such as an ArrayRef to RunArray<T>, panic’ing on failure.

as_string_array

Force downcast of an Array, such as an ArrayRef to StringArray, panicking on failure.

as_struct_array

Force downcast of an Array, such as an ArrayRef to StructArray, panicking on failure.

as_union_array

Force downcast of an Array, such as an ArrayRef to UnionArray, panicking on failure.

downcast_array

Downcasts a dyn Array to a concrete type