arrow_array::record_batch

Trait RecordBatchReader

pub trait RecordBatchReader: Iterator<Item = Result<RecordBatch, ArrowError>> {
    // Required method
    fn schema(&self) -> SchemaRef;
}

Trait for types that can read RecordBatch’s.

To create from an iterator, see RecordBatchIterator.

Required Methods

fn schema(&self) -> SchemaRef

Returns the schema of this RecordBatchReader.

Implementation of this trait should guarantee that all RecordBatch’s returned by this reader should have the same schema as returned from this method.